+++ application/modules/admin/views/index.php
@@ -42,15 +42,13 @@
           </div>
         </div>
         <div class="uk-grid uk-grid-small uk-grid-match uk-child-width-1-1 uk-child-width-1-2@s" data-uk-grid>
-          <?php foreach ($this->realm->getRealms()->result() as $charsMultiRealm):
-            $multiRealm = $this->realm->realmConnection($charsMultiRealm->username, $charsMultiRealm->password, $charsMultiRealm->hostname, $charsMultiRealm->char_database);
-          ?>
+          <?php foreach ($this->realm->get_realms() as $realm): ?>
           <div>
             <div class="uk-card uk-card-default">
               <div class="uk-card-header">
                 <div class="uk-grid uk-grid-small" data-uk-grid>
                   <div class="uk-width-auto">
-                    <h4 class="uk-h4"><span class="uk-margin-small-right"><i class="fas fa-server"></i></span>Realm - <?= $this->realm->getRealmName($charsMultiRealm->realmID); ?></h4>
+                    <h4 class="uk-h4"><span class="uk-margin-small-right"><i class="fas fa-server"></i></span>Realm - <?= $realm->name; ?></h4>
                   </div>
                   <div class="uk-width-expand uk-text-right">
                     <a href="<?= site_url('admin/realms'); ?>" class="uk-icon-button" uk-tooltip="<?= lang('admin_nav_realms'); ?>"><i class="fas fa-cog"></i></a>
@@ -61,17 +59,17 @@
                 <div class="uk-grid uk-grid-small uk-grid-divider uk-text-center" data-uk-grid>
                   <div class="uk-width-1-2">
                     <h5 class="uk-h5 uk-text-primary uk-text-bold uk-text-uppercase uk-margin-remove"><?= lang('info_alliance_players'); ?></h5>
-                    <h1 class="uk-h1 uk-margin-remove"><span class="blizzcms-count" data-from="0" data-to="<?= $this->realm->getCharactersOnlineAlliance($multiRealm); ?>" data-speed="2000" data-refresh-interval="50"></span></h1>
+                    <h1 class="uk-h1 uk-margin-remove"><span class="blizzcms-count" data-from="0" data-to="<?= $this->realm->count_online($realm->id, 'alliance'); ?>" data-speed="2000" data-refresh-interval="50"></span></h1>
                     <p class="uk-text-small uk-margin-remove"><?= lang('info_alliance_playing'); ?></p>
                   </div>
                   <div class="uk-width-1-2">
                     <h5 class="uk-h5 uk-text-danger uk-text-bold uk-text-uppercase uk-margin-remove"><?= lang('info_horde_players'); ?></h5>
-                    <h1 class="uk-h1 uk-margin-remove"><span class="blizzcms-count" data-from="0" data-to="<?= $this->realm->getCharactersOnlineHorde($multiRealm); ?>" data-speed="2000" data-refresh-interval="50"></span></h1>
+                    <h1 class="uk-h1 uk-margin-remove"><span class="blizzcms-count" data-from="0" data-to="<?= $this->realm->count_online($realm->id, 'horde'); ?>" data-speed="2000" data-refresh-interval="50"></span></h1>
                     <p class="uk-text-small uk-margin-remove"><?= lang('info_horde_playing'); ?></p>
                   </div>
                 </div>
                 <hr class="uk-divider-icon uk-margin-small">
-                <h6 class="uk-h6 uk-text-uppercase uk-margin-remove uk-text-center"><i class="fas fa-user-friends"></i> <span class="blizzcms-count uk-text-bold uk-text-warning" data-from="0" data-to="<?= $this->realm->getAllCharactersOnline($multiRealm); ?>" data-speed="2000" data-refresh-interval="50"></span> <?= lang('info_players_playing'); ?></h6>
+                <h6 class="uk-h6 uk-text-uppercase uk-margin-remove uk-text-center"><i class="fas fa-user-friends"></i> <span class="blizzcms-count uk-text-bold uk-text-warning" data-from="0" data-to="<?= $this->realm->count_online($realm->id); ?>" data-speed="2000" data-refresh-interval="50"></span> <?= lang('info_players_playing'); ?></h6>
               </div>
             </div>
           </div>