+++ application/modules/store/views/category.php
@@ -11,11 +11,11 @@
               </div>
               <ul class="uk-nav-default nav-store uk-nav-parent-icon" uk-nav>
                 <li><a href="<?= base_url('store'); ?>"><i class="fas fa-star"></i> <?= lang('store_top_items'); ?></a></li>
-                <?php foreach ($this->realm->getRealms()->result() as $MultiRealm): ?>
+                <?php foreach ($this->realm->get_realms() as $realm): ?>
                 <li class="uk-parent">
-                  <a href="javascript:void(0);"><i class="fas fa-server"></i> <?= $this->realm->getRealmName($MultiRealm->realmID); ?></a>
+                  <a href="javascript:void(0);"><i class="fas fa-server"></i> <?= $realm->name; ?></a>
                   <ul class="uk-nav-sub uk-nav-parent-icon" uk-nav>
-                    <?php foreach($this->store_model->getCategories($MultiRealm->realmID)->result() as $menulist): ?>
+                    <?php foreach($this->store_model->getCategories($realm->id)->result() as $menulist): ?>
                     <?php if($menulist->main == '2' && $menulist->father == '0'): ?>
                     <li class="uk-parent">
                       <a href="#"><?= $menulist->name ?></a>